fix(bhfindpatient): only expose api w/ flag set
This commit includes an additional check on the `$onInit` method for the
`bh-find-patient` component. This check ensures the API expose flag has
been set before executing the callback, this is inline with the optional
binding used by the component.

It also updates `vm` to `$ctrl` in favour of Angular component naming
standards.