fix(cashbox): user cashbox permission assignment
This commit introduces two checks on the server /users/:id/cashboxes
route that ensure that the correct data is provided before continuing,
this lets the client know if a request is badly formed in order to tell
the user.

It also updates the client controller to remove the default case where
it would send a badly formed request if no changes were made.