feat(vouchers): server voucher tools API end point
This commit introduces an API end point for the `correct` voucher tool.
This method is responsible for both reversing any journal transaction
and creating a generic voucher to replace it with corrected values. This
is achieved by calling the server controllers responsible for journal
transaction reversal and voucher creation and ensuring that the results
are cleaned up on failure.