fix(build): glyphicon path issue and corrupt css
This commit resolves two issues that were introduced by changes in the
build process:
1. The bootstrap less font paths are relative, that means the
bhima-bootstrap file must be located relative to the ../fonts folder.
This is moved.

2. Minifying the bootstrap CSS with out current minifier `cssnano` seems
to override some styles and animations causing issues on `ui-grid` row
validation. This has been reverted until it can be determined why this
minification causes this.