chore(bhCardList): rename component API; fix filtering issue
This commit renames the exposed binding on the bh-card-list component
to more clearly represent their purpose, favouring sort-name, sort-age
and sort-size over name, age and size respectively.

It also resolves an issue with unit tests failing on filtering due to an
issue with ng-model with a variable attribute, simply supporting
filtering all elements on all attributes resolved this issue.