refactor(patients): use async/await
This commit migrates the patients controller on the server to use
async/await where possible.  It also updates the patient reference
calculation to pull the values from the entity_map table.  Finally, it
makes sure that the entity_map table is kept up to date with any changes
made to the patient name.

Contributes to #2458.