fix(patients): fix update patient entity map
The patient entity map was clobbered when a patient's name was updated.
This led to an inability to sort the patient references.  I've addressed
this by updating the buggy code and re-running zRecomputeEntityMap() in
production.  Also, the sorting algorithm defaults to string sorting if
it can't figure out how to sort the numbers.

Closes #3483.