fix(reports): rewrite multiple account statement
This commit rewrites the multiple account statement report to make it
both more efficient and correct.  In the previous version, the opening
balance and cumsum calculation was off and identical for each account.
This uses the old tools from the single account statement to make it
line up correctly.  I've also removed and consolidated the db queries
into fewer queries.