feat(vouchers): add lookup of hrEntity to API
This commit adds the ability to create vouchers with entity_uuids by
specifying an hrEntity that is automatically linked via helper
functions.  Currently, if an hrEntity does not match, the API simply
ignores it.  In the long run, we might want to block the user if we
cannot find a match for the entity specified - this would require
flipping to a series of db.one()s or comparing unique array lengths.
db.one() would be less efficient (more DB calls), but arguably simpler
to read and maintain.