fix(renderers/pdf): rm unnecessary async code
This commit makes a few asynchronous code portions synchronous (since
they don't need to be async).  It also launches a browser instance PER
HTTP request.  In order to maintain a chromium process in the
background, we would need to be able to listen for any exit events and
close the chromium process.  However, best practice seems to be to start
up a new browser on each request.