refactor(cash): improve layout of cash cancel
The cash cancel modal has been revamped with a layout more resembling
the credit note modal of the invoice page.  Additionally, the confusing
code that referenced "invoice" (since it was copied and pasted) has been
removed to properly reference cash payments when they need to be
reference and invoices when they need to be referenced.

Additionally, the case of cautions are handled, so that we do not
display a list of invoices in this case.