fix(reports): handle unpaid balance errors
The Unpaid Invoices errors would throw SQL errors if there was no data
for the range (since that is what the Pivot() function does).  The
report has been updated to catch and gracefully handle certain data
errors so that it simply informs the user that there is no data for
that range.

Closes #3575.