feat(journal): allow more modifications to txns
We have implemented the ability to modify the transaction date of
transactions to allow users to change the date if it was a date error
that was made on a posted record.  Additionally, the transaction type
can also be modified.  Note that both of these options were available
for unposted records - these changes simply extend those options to
  posted records.