test(pdf): ensure template renders w/ barcode
The "dumb" noop() templating function in the PDF unit tests has been
replaced by the handlebars for faithfulness to the real-world use case.
The new mock HTML renderer allows us to test the added barcode feature
and ensure it works by rendering an HTML file with a barcode in it.