chore: add migration script generator
A migrations script generator has been added directly to repository and
to the npm scripts section of the package.json.  The idea is that it
will read the configuration found in the .env file and create a
migration script that can be run to upgrade the files to the next
version.