fix(cashflow): add gain/loss on exchange calc
The cashflow by service report doesn't take into account the gain/loss
on exchange since it is transparently handled by the system in the cash
module.  We've added in a final line to show the gain/loss on exchange
so that this information can be included in the final balance.