chore: update receipt css for pdfs
We've made a bunch of changes to our rendering engines.  We front-load
the asset inlining to the HTML renderer, since it permits us to use the
same paths everywhere instead of maintaining local rendering paths (full
paths on disk) and HTTP-accessible paths.

To properly render out POS receipts, a new stylesheet has been created
and put into a handlebars partial.  We have a single, central location
to alter the styles of our POS receipts.  This also allowed us to remove
the underline partial since it can be done in a stylesheet much more
easily.

The dimensions of the documents are now defined in CSS rather than in
the PDF renderer.  The only options passed in are "preferCSSPageSize" to
ensure that the templates are rendered according to their internal
stylesheets.