fix(app): do not crash on unCaughException
This reverses an earlier decision to crash the server
onUncaughtExceptions.  Unfortunately, we cannot catch wkhtmltopdf
exceptions and a simple missing <img src> will cause the server to
crash.  In particular, the enterprise logo will crash the server, and we
have no guarantees if this will occur in production.  Once we merge #3984
we may revisit this decision.