feat(reports): add file type detection
Adds file type detection to the reports archive page so that we are not
proposing to the user to print reports that are not printable (xlsx for
example).  It also adds features to provide the correct font awesome
icon depending on the extension to help a user determine which kind of
file they will be downloading.