feat(stock): allow exit by barcode
Implements the initial stock exit by scanning an invoice barcode.  In
addition to being able to attach a patient and/or an invoice, the user
is now able to scan an invoice barcode to look up the invoice in
question.

TODO - the barcode scan does not template in the values for each
inventory item like the bh-find-invoice component does.