fix(patients): remove all patient groups submits
Removing all patient groups from a patient didn't work previously.  This
was because we had an invalid check in the submit method that would
never send the request to the server.  Now, we check that either there
are groups to submit or the user is removing all groups.

We also ensure that if the user is removing all groups, the checkboxes
are cleared and vice versa.