fix(tests): various patches
We have a number of tests that still depend on control flow. I've
cleaned up several instances where async/await was not used or not used
properly.

Improves the targeting of the invoicing fees in the debtor groups test.

Not sure why we have to do this all of a sudden, but to get past the
posting journal, you need to ignoreSynchronisation.

Rewrites the inventory groups configuration modal to follow better
practices and hopefully eliminate the errors found on Travis with the
latest Chrome.