refactor(reports): rename stock_sheet
Adds in the proper name of the Stock Sheet (Fiche de Stock) in English
and updates all paths to reflect this.  Previously, we had multiple
confusing names (inventory_file on the client, inventory_report on
server).  Now, it's all `stock_sheet`.

Also updates the parameter parsing to in line with most other reports by
removing JSON.parse() from the URL and making the URL readable.

Finally, we have added SUMs to the bottom of the stock_sheet.

Closes #4386.