refactor: make dropdowns uniform
Goes through all dropdowns in the application and makes them uniform by
aligning them to the right (`.text-right`) and also making sure they
don't overflow but instead use `text-ellipsis` to overflow.  I've also
put headers on the dropdown menus that do not have headers and expanded
the text of those where it would be more useful.

Closes #4586.