fix(stock): activate cancel button on depot modal
Ensures that the "cancel" button works on the depot selection modal.
This means that users can always cancel an operation and it will not
leave them in an undefined state.  If there is no default depot, it
returns them to the previous state.  If there is a defined depot, it
returns them to the current module.

Partially addresses #4776.