fix(stock): movements should group by is_exit
Stock movements are not uniquely identified by their document_uuid.
This is documented in #4083.  However, the tests made this assumption
because MySQL 5.7 had a stable sort on the GROUP BY.  In MySQL 8, this
is no longer stable.

To solve this issue, the grouping on document_uuid now includes the
is_exit column as well.