feat(users): add "last_login" functionality
This commit adds the functionality to actually display the last login
date of the user.  This column had been playing the role of "updated_at"
rather than "last_login" since it was only changed automatically when
user information had been changed.  Now, the last_login it updated
whenever a user logs in.

I've also reset all values of the last_login to NULL in the migration
script.

Closes #5216.