fix(stock): make "at risk of expiration" flag work
This fix restores the functionality of the "at risk of expiration" flag
on the articles in stock registry.  It does not affect the API, just the
internal calculations.

In addition, we also changed the behavior of the articles in stock
action dropdown menu so the link to the stock lots now filters on the
depot and time period in question.

Closes #5422.