feat(stock): add minimum delay to stock settings
Adds a "minimum delay" setting to stock management to ensure that we are
able to detect stock outs on time, even if the purchase data is messy.
Previously, we relied on purchasing data solely for the delay, and users
who put in a purchase order, then immediately delivered it, shoved the
delay closer to 0.  This caused us to not warn users in time of stock
outs.  The user can now configure the minimum number of months in
decimal form on the stock settings page.

Closes #5407.