fix(stock): count initial days as stock out
This commit changes the CMM calculation to ensure that _head_days is
added to the _sum_stock_out_day variable when there are no previous
records.  This can happen when a depot has just newly opened and the
date range used for calculating the CMM is greater than the number of
days the depot has been operating.

Closes #5482.