feat(purchase): update purchase orders
This commit implements the ability to update non-confirmed purchase
orders.  If a purchase order has been confirmed, the only thing that
should be updated is the status (canceled).  This is yet to be
discussed.

To modify a purchase order, one must go to the purchase registry and
select "edit" on the purchase orders that is non yet confirmed.  This
will take you to the regular purchase order creation page with all
values filled out.

Closes #5489.