fix(purchase): allow canceling orders
This commit allows purchase orders to be canceled.  It also adds a
updated_at column to the purchase table so that we can keep track of
when edits happen.  Finally, canceling a purchase order is a sink - we
can't undo a cancel; however, you can still cancel an order that is
confirmed.

Closes #5489.