feat: remove mailgun and add nodemailer
This commit removes mailgun-js which is unmaintained and has been
unusable for us for a while due to shutting down their free tier.
Instead, we replace it with nodemailer which lets you set your own SMTP
credentials as environmental variables.  In the long run, we should
probably provide these as configuration settings a user can set for the
application.

Closes #5290.