fix(stock): do not hang on requisition search
This commit prevents the stock exit to requisition from hanging.  It
does this in two ways - first, it removes the typo that is causing the
infinite hang.  Second, it causes the requisition select to be disabled
if the depot isn't selected.  This should show users that they have a
problem and drive them to look up the depot.

Generally, the best way would be to improve the hasRequisitionVoucher
code, but it can't be done quickly.  Hence this quick fix.

Closes #5642.