fix(xlsx): do not crash on invalid date
If a date isn't valid, the XLSX renderer would just crash, dumping JSON
to the client.  By checking if the date is valid prior to passing it to
XLSX, we prevent a crash by rendering the invalid date as a string so
users can inspect the date and change it.

Closes #5666.