fix(stock): properly exchange stock value
The stock value report was updated to include exchanged values so that
the report could be viewed in $USD at Vanga.  However, the exchange rate
was incorrectly applied, causing bugs in the Weighted Average Cost
algorithm.  To fix this, we apply only the rate to the `unit_cost` and
carry that forward via sums.