feat(invoices): automatically allocate costs
This commit implements the automatic cost allocation during invoicing.
The logic is that any income/expense account that is associated with a
cost center is assigned that cost center when writing to the
posting_journal.  If the account is not assigned, we default to the
service's cost center.

This applies to all invoice_items, invoicing_fees, and
invoice_subsidies.

Closes #5855.