fix(stock): improve Stock Value Report
This commit fixes various components of the stock value report cited in
issue #6150.  Notably, the query now uses the stock_movement_status
table to pull in the quantities in stock for a given depot and loads the
stock_value table to get the value of the items in stock.  This speeds
the query up signficantly.

Closes #6150.