chore: initialize mysql database in docker
The mysql dockerfile has been pinned at 8.3 to avoid errors in the
default authentication plugin.  See
https://github.com/mysqljs/mysql/pull/2233 for more information.  It
might be prudent to switch to mysql2 instead of mysqljs for future
release.

I've renamed all the server model files to have a numeric prefix to
allow execution in order.  The 99-debug.sql file crashes, but it is not
necessary for the functioning of the BHIMA server.

You can now type `docker-compose up` and be taken to the BHIMA
installation page on localhost.  Environmental variables are set in the
`.env` file as usual.