deps: upgrade to gulp v5
This commit upgrades gulp to version 5.  The latest gulp revision
changes the streaming library which defaulted src() to UTF-8 encoding
for all streams.  This worked seamlessly for everything except fonts,
which became terribly mangled.  The fix is to set encoding:false on
gulp src().

I've also upgrade a few other dependencies and shifted some of the
developmental dependencies from the "dependences" to "devDependencies".