feat(reports): use a datauri for enterprise logo
This commit uses a datauri encoding for the enterprise logo.  It also
fixes the logo rendering issue called by the gulp v5 update which
created encoding issues during the build step.

Secondarily, it ensures the enterprise logo exists before using it, as
well as including a new partial for point-of-sale receipts to avoid
repeat code usage.

This should fix the broken enterprise logo reported by @lomamech in
 #7679.

Partially addresses #7679.