feat(payroll): add column w/ number of configured employees
This commit adds a column to the employee configuration page in the
human resources folder to indicate the number of configured employees
out of the total number of employees.

Note: the total number of employees is computed at the time the request
is made, not at the time that the configuration was created.   Thus,
employees hired after the fact will appear on the list, and may lead to
confusion.

Closes #7755.